Move non-curve-specific code from document.h and add laplace-curve construction from a directory of files rather than a single file
Pretty much every program that deals with directories of files now
reads a file with a listing of the directory with the files to process,
rather than read the directory itself. This is done in all steps of the
curve pipeline (starting from the construction of the LDA input file)
so we can ensure we posteriorly read the files by the order they appear
in the LDA matrix (this is important when we fetch the class labels
from the file names in the classification step). Since we need this
behavior in this case and standardizing on this is not a bad thing, we
employ this directory listing method everywhere where the alternative
is to read the directory directly.