Fix curve composition issues. Add curve averaging program
Objects inside a document were not being properly moved from
when the document changed scope (such as putting it inside
a vector, for example), which would cause a trove of use
after free errors whenever we tried to evaluate a curve in a
new scope. This, of course, also caused problems for composite
curves, whose functionality depends on the existence of other
documents with properly founded (matrix-based) curve functions.
This meant that any document created from composing two other
documents was ill-formed.